chore: centralized event subscribers
